==> scripts/01_dump_to_pd.ipynb <==
## Bibliotheken:
from bs4 import BeautifulSoup
from lxml import etree as ET
import pandas as pd
import unicodedata
from tqdm import tqdm
import polars as pl

#Einlesen des Dumps in ElementTree:

xml = open("ohs_03_2024.xml", "r", encoding="utf-8")
tree = ET.parse(xml) 
root = tree.getroot() 

records = list(root)
total_records = len(records)  

#Ausgabe der Gesamtzahl der in der Datei vorhandenen Records:
print("Gesamtzahl bibliographischer Einträge: ",total_records)

#Function to get actual data: 
def parse_record(xml):
      
    ns = {"marc":"http://www.loc.gov/MARC21/slim"}
        
    #IDN:      
    idn = xml.findall("marc:controlfield[@tag = '001']", namespaces=ns)
    try:
        idn = idn[0].text
    except:
        idn = 'fail'
        
    #008:      
    c008 = xml.findall("marc:controlfield[@tag = '008']", namespaces=ns)
    try:
        c008 = c008[0].text
    except:
        c008 = 'fail'
        
        
    #Language: 
    lang = xml.findall("marc:datafield[@tag = '041']/marc:subfield[@code = 'a']", namespaces=ns)
    try:
        lang = lang[0].text
    except:
        lang = 'fail'
        
    #Creator: 
    creator = xml.findall("marc:datafield[@tag = '100']/marc:subfield[@code = 'a']", namespaces=ns)
    creator_body = xml.findall("marc:datafield[@tag = '110']/marc:subfield[@code = 'a']", namespaces=ns)
    if creator:
        creator = creator[0].text
    elif creator_body:
        creator = creator_body[0].text
    else:
        creator = "fail"    
        
        
    #Title: 
    maintitle = xml.findall("marc:datafield[@tag = '245']/marc:subfield[@code = 'a']", namespaces=ns)
    subtitle = xml.findall("marc:datafield[@tag = '245']/marc:subfield[@code = 'b']", namespaces=ns)
    if maintitle:
        maintitle = maintitle[0].text
        if subtitle: 
            title = maintitle + " : " + subtitle[0].text
        else:
            title = maintitle
    else:
        title = "fail"    
    
    
    #Place of publication: 
    place = xml.findall("marc:datafield[@tag = '264']/marc:subfield[@code = 'a']", namespaces=ns)
    diss = xml.findall("marc:datafield[@tag = '502']/marc:subfield[@code = 'a']", namespaces=ns)
    diss2 = xml.findall("marc:datafield[@tag = '502']/marc:subfield[@code = 'c']", namespaces=ns)
    if place:
        place = place[0].text
    elif diss:
        place = diss[0].text
    elif diss2:
        place = diss2[0].text
    else:
        place = "N/A"    
        
        
    #Year of publication: 
    year = xml.findall("marc:datafield[@tag = '264']/marc:subfield[@code = 'c']", namespaces=ns)
    diss = xml.findall("marc:datafield[@tag = '502']/marc:subfield[@code = 'a']", namespaces=ns)
    c008_2 = xml.findall("marc:controlfield[@tag = '008']", namespaces=ns)
    if year:
        year = year[0].text
    elif diss:
        year = diss[0].text
    elif c008_2:
        year = c008_2[0].text
        year = year[7:11] + " - 008"
    else:
        year = "0"    
    
         
    #notes: 
    notes = xml.findall("marc:datafield[@tag = '500']/marc:subfield[@code = 'a']", namespaces=ns)
    if notes:
        notes = notes[0].text
    else:
        notes = "N/A" 
    
    #diss: 
    diss0 = xml.findall("marc:datafield[@tag = '502']/marc:subfield[@code = 'a']", namespaces=ns)
    diss1 = xml.findall("marc:datafield[@tag = '502']/marc:subfield[@code = 'b']", namespaces=ns) #Diss
    diss2 = xml.findall("marc:datafield[@tag = '502']/marc:subfield[@code = 'c']", namespaces=ns) #Uni
    diss3 = xml.findall("marc:datafield[@tag = '502']/marc:subfield[@code = 'd']", namespaces=ns) #Jahr
    if diss0:
        diss0 = diss[0].text
    else:
        diss0 = "none"
        
    if diss1:
        dissdiss = diss1[0].text
    else:
        dissdiss = "none"  
        
    if diss2:
        dissuni = diss2[0].text
    else:
        dissuni = "none"   
    
    if diss3:
        dissyear = diss3[0].text
    else:
        dissyear = "none"  
         
    
    if diss0 != "none": 
        diss = diss0
    elif (dissdiss != "none") & (dissuni != "none") & (dissyear != "none"):
        diss = dissdiss + " " + dissuni + " " + dissyear
    elif (dissdiss != "none") & (dissuni != "none"):
        diss = dissdiss + " " + dissuni
    elif dissdiss != "none":
        diss = dissdiss
    else:
        diss = "not found"   
    
    #DDC: 
    ddc3 = xml.findall("marc:datafield[@tag = '082']/marc:subfield[@code = 'a']", namespaces=ns)
    if ddc3:
        ddc3 = ddc3[0].text
    else:
        ddc3 = "N/A"   
        
        
    #DDC weitere: 
    ddc1 = xml.findall("marc:datafield[@tag = '083']/marc:subfield[@code = 'a']", namespaces=ns)
    if ddc1:
        ddc1 = ddc1[0].text
    else:
        ddc1 = "N/A" 
        
        
    #DDC?: 
    ddc2 = xml.findall("marc:datafield[@tag = '083']/marc:subfield[@code = 'q']", namespaces=ns)
    if ddc2:
        ddc2 = ddc2[0].text
    else:
        ddc2 = "N/A" 
    
 
    #ddc_prio
    if ddc1: 
        ddc_prio = ddc1
    elif ddc2:
        ddc_prio = ddc2
    elif ddc3:
        ddc_prio = ddc3
    else:
        ddc_prio = "fail"
        

    #Sachgruppe: 
    sw = xml.findall("marc:datafield[@tag = '084']/marc:subfield[@code = 'a']", namespaces=ns)
    if sw:
        sw = sw[0].text
    else:
        sw = "N/A"    
        
    
    #Zusammenführen:
    gathered = {'ID':idn, 'Creator':creator, 'Title':title, 'lang':lang, 'Place':place, "Year": year, 'DDC':ddc_prio, 'DDC1':ddc1, 'DDC2':ddc2, 
                'DDC-DE?':ddc3, 'Sachgruppe':sw, 'diss':diss, 'notes':notes, '008':c008}
    return gathered

#extract relevant data for app from dump: 

pbar = tqdm(total=total_records)

result = []
for item in records:
    result.append(parse_record(item))
    pbar.update()


#Convert to Dataframe and save as .h5: 

#df = pl.DataFrame(result)
#df.write_parquet("data/online_diss_03-2024.parquet")
df = pd.DataFrame(result)
df.to_hdf('data/online_diss_03-2024.h5', key='df')  

df
