==> scripts/02_clean_data.ipynb <==
#Bibliotheken:
from bs4 import BeautifulSoup
from lxml import etree as ET
import pandas as pd
import unicodedata
import polars as pl

#Read df from file:
dump =  pl.read_parquet('data/online_diss_03-2024.parquet') 
df = dump.to_pandas()




#df =  pd.read_hdf('data/online-diss_03-2024.h5', 'df') 
find_none = df.loc[df['diss'] == "not found"]

#df =  pl.read_parquet('data/online_diss_03-2024.parquet') 
#find_none = df.filter(pl.col('diss') == 'not found')
no_thesis = (len(find_none))
print("Anzahl der Datensätze, die keinen Hochschulschriftenvermerk enthalten: ", no_thesis)

print(len(df))

#Drop all records that don't have a dissertation statement:
df['diss'] = df['diss'].replace(["not found"], [None])
df = df.dropna(subset=['diss'])
df

df.filter(pl.col('Place') == 'N/A')

#Now clean columns "Place" and "Year":

df['Place'] = df['Place'].str.replace(',.*', '', regex=True) #For records with disseratation statement in "Place" keep string until first comma 
df['Place'] = df['Place'].str.replace("Zugl.: ", "")
df['Year'] = df['Year'].str.extract('(\d+)')
df


#df.to_hdf('data/online-diss-cleaned_03-2024.h5', key='df')

cleaned_dump = pl.from_pandas(df)

cleaned_dump.write_parquet("data/online_diss_cleaned_03-2024.parquet")
