==> scripts/03_create_pub_years.ipynb <==
from bs4 import BeautifulSoup
from lxml import etree as ET
import pandas as pd
import unicodedata
import matplotlib.pyplot as plt 
import plotly.express as px
import numpy as np
import polars as pl

dump = pl.read_parquet('data/online_diss_cleaned_03-2024.parquet') 
data = dump.to_pandas()
#data = pd.read_csv('data/check-data-csv.csv')
#data

#Get missing years from 008: 
data['Year'] = data['Year'].fillna(data['008'].str[7:11])

# Filter entries that don't have a publication date and reduce timespan: 

data1 = data[data['Year'].notna()]
data1.query("Year.str.len() == 4")

data1 = data1.astype({'Year':'int'})
data1 = data1[(data1['Year'] >= 1990) & (data1['Year'] <= 2023)]
data2 = data1[data1['Year'] >= 1990]

print("Data: ", len(data))
print("Data 1990-2023: ", len(data1))
print("Data >= 1990: ", len(data2))

#create new dataframe that only includes the publication year and value count: 
s = data1['Year'].value_counts().sort_index()

new = pd.DataFrame(s)
new = new.rename(columns={"Year": "count"})
new['years'] = new.index
new

#check data: 

df2 = new['count'].sum()
df2

no_data = data[data['Year'].isnull()]
no_data

data1.loc[data1["Year"] == "unknown"]

new.to_json("data/diss_years.json", index=True)
