==> scripts/04_create_ddc.ipynb <==
from bs4 import BeautifulSoup
from lxml import etree as ET
import pandas as pd
import unicodedata
import matplotlib.pyplot as plt 
import plotly.express as px
import numpy as np
import polars as pl

dump = pl.read_parquet('data/online_diss_cleaned_03-2024.parquet') 
data = dump.to_pandas()
#data = pd.read_csv('data/check-data-csv.csv')
#data

# Copy DDC3 to DDC where DDC = N/A: 
data_ddc = data.rename(columns={"DDC-DE?":"DDC3"})
data_ddc["DDC"] = np.where((data_ddc.DDC == "N/A"), data_ddc.DDC3, data_ddc.DDC)
data_ddc["DDC"] = np.where((data_ddc.DDC == "N/A"), data_ddc.Sachgruppe, data_ddc.DDC)
data_ddc

#data_ddc.query('DDC == Sachgruppe')
data_ddc.query('DDC == "N/A"')
#data_ddc.query('DDC.str.len() < 3')

no_ddc = data_ddc.loc[data_ddc['DDC'] == "N/A"]
print("Datensätze ohne DDC: ", len(no_ddc))

# Load mapping data: 
mapping = pd.read_excel("data/DDC_SGT_Mapping.xlsx", converters={'Sachgruppe':str, 'Zuordnung':str})
mapping.keys()


merged = pd.merge(data_ddc, mapping, on=["Sachgruppe"], how="left") #Merge both tables
#merged.query('DDC.str.len() < 3')

print(len(merged))

#show entrys with DDC < 3:
#merged.loc[merged['DDC'].str.len() < 3, 'DDC']

merged.loc[merged['DDC'].str.len() < 3, 'DDC'] = merged['Zuordnung'] #Mappe Zuordnung to DDC where DDC < 3 characters
print(len(merged))

#print(merged.loc[20342])
#merged.query('Sachgruppe.str.len() < 3')
#merged.query('ID == "1015899285"')

merged.loc[merged['DDC'] != "N/A",]

#merged.to_csv("data/stats.csv")

mainddc = merged['DDC'].str[:3]  # create list of 1. ddc entry (and shorten DDC to first 3)
data2 = pd.DataFrame(mainddc)  # convert to dataframe
data2["count"] = data2.groupby("DDC")["DDC"].transform('count') # add column for value_count of DDC-number
data2

data2 = data2.drop_duplicates("DDC") # drop duplicate entries based on DDC-number
data2 = data2.rename(columns={"DDC": "DDCthird"})
print(len(data2))

data2.query('DDCthird.isnull()' or 'DDCthird == "N/A"')

data2 = data2[data2.DDCthird != "N/A"]
data2 = data2[data2.DDCthird.notnull()]
data2

find_start = data2[data2["DDCthird"].str.startswith('2')]
df1 = find_start['count'].sum()
print(df1)

ddc = pd.read_excel('data/DDC_complete.xlsx', converters={'DDCthird':str, 'DDCsecond':str, 'Parent_no':str}) #Read DDC file
#ddc

df = pd.merge(ddc, data2, on=['DDCthird'], how="left")
#df = df.drop_duplicates("DDCthird")
df['count'] = df['count'].fillna(0)

#df

df.to_json("data/diss_ddc.json")
