==> scripts/05_create_geodata_pubplace.ipynb <==
from bs4 import BeautifulSoup
from lxml import etree as ET
import pandas as pd
import unicodedata
import matplotlib.pyplot as plt 
import plotly.express as px
import numpy as np
import requests
from bs4 import BeautifulSoup as soup
import polars as pl
#from arcgis import GIS
#import geopy
#from geopy.geocoders import Nominatim

#from arcgis.features import SpatialDataFrame
#from arcgis.features import GeoAccessor
#from arcgis.gis import GIS

dump = pl.read_parquet('data/online_diss_cleaned_03-2024.parquet') 
data = dump.to_pandas()
#data = pd.read_csv('data/check-data-csv.csv')
data

data.loc[data['Place'] == "N/A"]

##Um Probleme mit Sonderzeichen beim Geocodieren zu vermeiden, normalisieren wir noch einmal das Dataframe mit Hilfe einer Funktion: 

def normalize(text): 
    new = unicodedata.normalize('NFC', text)
    return new 

data['Place'] = data['Place'].apply(normalize)
data['Place'] = data['Place'].str.strip('[]')
#df['Place'].replace("Karlsruher Inst. für Technologie", "Karlsruhe", inplace=True)
#df['Place'].replace("Hagen. Fernuniv.", "Hagen", inplace=True)

df = data[["ID", "Title", "Place"]]
df = df.drop_duplicates("Place")
#df.to_csv("places_to_reconcile.csv")
df

df.loc[df['Place'] == "N/A"]

### Im nächsten Schritt wurde mit Hilfe von OpenRefine und der Lobid-Reconciliation-API der Normdatensatz in der GND für jedes Geografikum bestimmt und diesem zugeordnet. Im folgenden wird die um GND-ID erweiterte Tabelle geladen und auf die ursprüngliche Liste aller Publikationen gemappt. 

gnd_places = pd.read_excel("data/places-with-gnd.xlsx")
#gnd_places.loc[gnd_places['Place'] == "N/A"]
gnd_places

new_places = df.merge(gnd_places, on='Place', indicator=True)
new_places.loc[new_places['_merge'] != "both"]

new_places.loc[new_places['Abgleich'] != new_places['GND']]

places = gnd_places['GND-ID'].to_list()
print("Anzahl GND-IDs in Places: ", len(places))

## Jetzt noch Geokoordinaten über SRU der GND abfragen und hinzufügen: 

def gnb_sru(gndid):
    
    base_url = "https://services.dnb.de/sru/authorities"
    params = {'recordSchema' : 'MARC21-xml',
          'operation': 'searchRetrieve',
          'version': '1.1',
          'maximumRecords': '100',
          'query': 'nid="' + str(gndid) + '"'
         }
    
    r = requests.get(base_url, params=params)
    xml = soup(r.content, features="xml")
    records = xml.find_all('record', {'type':'Authority'})
    #print("Treffer: ", len(records))
    treffer = xml.find_all("numberOfRecords")[0].text
    #print("number of records: ", treffer)
    
    if len(records) == 1: 
        
        #print("Es geht weiter")
        for record in records:
            content = record
    
        ns = {"marc":"http://www.loc.gov/MARC21/slim"}
        xml1 = ET.fromstring(unicodedata.normalize("NFC", str(content)))
        long = xml1.xpath("marc:datafield[@tag = '034']/marc:subfield[@code = 'd']", namespaces=ns)
        lat = xml1.xpath("marc:datafield[@tag = '034']/marc:subfield[@code = 'f']", namespaces=ns)
        for elem in long:
            if "." in elem.text: 
                long = elem.text
                if long.startswith("E"):
                    long = long.replace("E", "")
                    long = long.lstrip('0') 
                elif long.startswith("W"):
                    long = long.replace("W", "")
                    long = long.lstrip('0') 
                    long = "-" + long
                #print(long)
        for elem in lat:
            if "." in elem.text: 
                lat = elem.text
                if lat.startswith("N"):
                    lat = lat.replace("N", "")
                    lat = lat.lstrip('0') 
                elif lat.startswith("S"):
                    lat = lat.replace("S", "")
                    lat = lat.lstrip('0') 
                    lat = "-" + lat
                #print(lat)
        
        coordinates = {'GND-ID': gndid, 'lat': lat, 'long': long}
        return coordinates
             
    else:
        
        print("ERROR - ", gndid)
        coordinates = {'GND-ID': gndid, 'lat': ' ', 'long': ' '}
        return coordinates
        

myquery = [gnb_sru(place) for place in places]


df2 = pd.DataFrame(myquery)
df2

new = pd.merge(gnd_places, df2, on="GND-ID", how="left")
combined = new[["Place", "Abgleich", "GND", "GND-ID", "lat", "long"]].copy()
combined = combined.drop_duplicates("Place")
combined

combined.to_excel("data/geoplaces_coordinates_from_gnd.xlsx")

complete = pd.merge(data, combined, on="Place", how ="left")

complete["count"] = complete.groupby('GND-ID')['GND-ID'].transform('count')
complete

places = complete[["Place", "Abgleich", "GND", "GND-ID", "lat", "long", "count"]]
places = places.drop_duplicates("GND-ID")
places = places[places.lat != "[]"]
places = places[places.lat != ""]

places['lat'] = places['lat'].str.replace(r'^[0]*', '', regex=True)
places['long'] = places['long'].str.replace(r'^[0]*', '', regex=True)
#places['count'] = places['count'].str.replace(".0", "")
places

places = places[places.lat != ""]
places = places[(places['lat'].notnull())]
places

places.to_excel("data/geoplaces.xlsx")

places.to_json("data/geoplaces.json")

## OLD

## Connect to Geolocator:
geolocator = Nominatim(timeout=10, user_agent = "myGeolocator")

#Add new colum 'gcode' that contains translated Ort-value to geolocation: 
#WARNING: THIS MAY TAKE A WHILE

dfplaces['gcode'] = dfplaces.Place.apply(geolocator.geocode)
dfplaces

## Find entries where gcode didn't find a match: 

dfmissmatch = dfplaces[dfplaces['gcode'].isnull()]
dfmissmatch

#Split into long and lat: 
dfplaces['lat'] = [g.latitude for g in dfplaces.gcode]
dfplaces['lon'] = [g.longitude for g in dfplaces.gcode]

dfmerge = pd.merge(df, dfplaces, on='Place', how='left')
dfmerge["count"] = dfmerge.groupby('Place')['Place'].transform('count')
#dfmerge.to_csv("check_places.csv")
dfnew = dfmerge.drop_duplicates(subset=['Place'])
dfnew

dfnew.to_json("data/geoplaces.json")

test = pd.read_json("data/geoplaces.json")

test
