==> scripts/07_get_language.ipynb <==
import pandas as pd

data = pd.read_hdf('data/online-diss-cleaned_11-2023.h5', 'df')
#data = pd.read_csv('data/check-data-csv.csv')
data

stats = data[["ID", "lang", "Year"]]
#stats[stats['Year'].isnull()]
#stats

lang = stats.groupby(['lang']).size().reset_index(name='counts')
sprachen = pd.read_excel("data/Sprachen.xlsx")
sprachen


lang = pd.merge(lang, sprachen, on=['lang'], how="left")
lang

lang = lang[lang.lang != "und"]
lang = lang[lang.lang != "fail"]
lang

lang.to_json("data/diss_lang.json", index=True)

all_lang = (lang['counts']).sum()
print(all_lang)
